Moved logic into the subclass.
